Error check copying include file
Check if copying php.ini fails, and if so terminate script.

+++ nextcloud-jail.sh
@@ -375,7 +375,11 @@ if [ $SELFSIGNED_CERT -eq 1 ]; then
 fi
 
 # Copy and edit pre-written config files
-iocage exec "${JAIL_NAME}" cp -f /mnt/includes/php.ini /usr/local/etc/php.ini
+if ! iocage exec "${JAIL_NAME}" cp -f /mnt/includes/php.ini /usr/local/etc/php.ini
+then
+	echo "Failed to copy php.ini"
+	exit 1
+fi
 iocage exec "${JAIL_NAME}" cp -f /mnt/includes/redis.conf /usr/local/etc/redis.conf
 iocage exec "${JAIL_NAME}" cp -f /mnt/includes/www.conf /usr/local/etc/php-fpm.d/
 if [ $STANDALONE_CERT -eq 1 ] || [ $DNS_CERT -eq 1 ]; then
